Avoiding the confusion with the gpg module
It turns out that the pip package gnupg is not the same than the rpm package
python-gnupg, and that the latter has a pip package named python-gnupg.

That and the fact that the modules were 'compatibe' until now, created the
illusion that the tests were being done with the same packages than where it
would run latter.

Change-Id: I7d3896ca92edcb773f34854a27ec512c610aeb4d
Signed-off-by: David Caro <dcaroest@redhat.com>