RPM: move from createrepo to createrepo_c
createrepo_c is available on both CentOS and Fedora.
It is faster than createrepo and doesn't depend on yum so
it doesn't pull in a deprecated package manager on Fedora.

If createrepo_c is not available, repoman will fallback
on createrepo. This applies to manual installations from
source code since the rpm will require createrepo_c.

Change-Id: I960e4f2e8027c53f2b441fd2efde329857f9d88b
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>