Removing defaults for to_all_distros packages.
Repoman had hardcoded list of packages that should be copied to
all distros by default. However, sometimes the list of all distros is
not defined and it produces a problem for default flows. Since if the
package from the list has a distro defined, repoman will ignore it
and think that it is still distroless that is not good to happen by default.

Users are still able to define to_all_distros packages in configuration 
if needed. This way we keep repoman general and isolate those settings
in configuration of specific setups.

Change-Id: Iead55c8eaf66696abfa9125c7e2e2edf90513e3e
Signed-off-by: Anton Marchukov <amarchuk@redhat.com>