Added Dockerfile for the development image.
The development docker image is based on the CentOS 7 image and
includes all dependencies required to build repoman, run test
suits and generate the docs.

The image is intended to provide reproducible environment independed
from system used for development.

README.rst is updated with the instructions on how to use the image.

Change-Id: Ia67cdb8d65329a29dc04306182fa911425b07f75
Signed-off-by: Anton Marchukov <anton@martchukov.com>