Change only-missing filter to use full_name.
If only-missing filter is not using artifacts "full_name" it will consider
binary RPM to be the same as its SRPM because regular "name" does not
include srpm flag. Due to this it will take only one of them from the
repo where SRPM is provided along with the binary one.

full_name is now a full property to make sure it is part of the contract.

Change-Id: I29891132e6af4dc2af9e9b4d8b2e546504dcc394
Signed-off-by: Anton Marchukov <amarchuk@redhat.com>