RPM: delete signatures before signing
Remove any existing signature from the rpm before signing it.
This is needed because if a signature already exist, even whith our
signature, when installing it yum raise an error like:
The GPG keys listed for the "oVirt 4.2 Pre-Release" repository are
already installed but they are not correct for this package.

In order to detect the wrong signature instead of re-signing all
the packages, added logic for retrieving the signing key used
for the rpm signature.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1390839
Change-Id: I4453dd2cbd2d780539b3324dc1dd13a3288f09fb
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>