Added a repo container builder image
Added files needed to build a container image that contains `repoman`
and an HTTP server and can be used as an s2i builder image to create
container image that expose package repositories that were generated
using `repoman` sources list files.

This image allows one to use OpenShift to create "cloud native" package
servers that are comprised of a set of interlinked containers.

Also added a debug mode to the builder image to enable debugging builds.

Change-Id: I30a20d0d1ad85cf8c7f60cf6f7aa07a7fad1c573
Signed-off-by: Barak Korren <bkorren@redhat.com>