py3: Fix broken 'ConfigParser' usage
Looking at:

 Python 2.7.17 (default, Oct 20 2019, 00:00:00)
 >>> import ConfigParser
 >>> help(ConfigParser.SafeConfigParser.get)

we can see that the third argument passed to 'SafeConfigParser.get'
is 'raw', which should be boolean. In our code we use an empty string.
On py3 'raw' is a keyword-only argument so this call ends up with error.
This is probably a simple mistake and everything works without it.

Change-Id: I63aa465a9512704ce55618b0ff300f203dfe7b09
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>